Add configuration for ING (NL) savings account specific exports
The ING in the Netherlands provides options for 2 CSV exports; an export for the main account (see defaults.json) and one for an associated savings account.  
The CSV files of each export are supplied in a different format and are not interchangeable. A separate configuration file is required for savings account exports.  

This adds a configuration file for the savings account export variant.